Electron Conversion
Converted the whole app to use electron. You will need to install Node.js and might need to set the path variables.

After Node.js is installed, in the vs code terminal run: npm i -D electron
This will install everything for electron. This is over 100MB so DO NOT push it to github as it will make it so we can't make anymore pushes. It should be in the git ignore file already.

To run the desktop app, in vs code console run: npm start