Auth0 Setup
This is the base setup/logic to authenticate users with Auth0. This resulted in some modifications to how the app looks and where files are located. The home.html will need to be updated to look like the rest of the app. To get a better understanding of the changes, read this article: https://auth0.com/blog/securing-electron-applications-with-openid-connect-and-oauth-2/