User creation
Users are now created both in auth0 and truenas. If an error is encountered in the creation of a user for truenas it will log the user out of the application and attempt the process again. The next step is to set the max volume size for the user and then turn there home directory into a share drive. Proper error information for the user will need to be added later, most likely a popup window that will tell them the error code and to contact support if this keeps happening.

There was an issue with getting the api to work and I needed to rebuild the project. I left some files in different locations as to not risk breaking anything. This will be the new file scheme.