Custom title bar with full functionality, a start on a hamburger menu, and drive mounting fixed
Custom title bar that has min, max, close, and drag functionality. Started on a hamburger menu, it currently folds out with no way to fold back in and the styling is really rough, but the buttons work. "Fixed" drive mounting, it is a password issue with a temporary fix that works but isn't the best.